refactor(composables): lift useReorderDrag out of deck domain
It's geometry-agnostic already and about to gain a second caller
(dashboard deck reordering), so it belongs as a cross-feature
primitive rather than under views/deck/card-editor.
